+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - Hide user menu and disable links to home page on approval form
 - Replaced Cypress with TestCafe
+- Default meta description
 
 ## [1.0.1] - 2020-08-31
 ### Fixed